Fix DHT bootstraping
Fixes #1

Fixes bad translation of the DHT nodes' ports to big endian

+++ settings.h
@@ -37,7 +37,7 @@ public:
         QString name;
         QString userId;
         QString address;
-        int port;
+        quint16 port;
     };
 
     const QList<DhtServer>& getDhtServerList() const;
